refactor: smaller engine interfaces
