+++ internal/engine/engine_reader.go
@@ -4,10 +4,6 @@ import (
 	"context"
 	"fmt"
 	"iter"
-	"strings"
-	"time"
-
-	"stackit.dev/stackit/internal/git"
 )
 
 // AllBranches returns all branches
@@ -49,48 +45,6 @@ func (e *engineImpl) ValidateOnBranch() (string, error) {
 		return "", fmt.Errorf("not on a branch")
 	}
 	return currentBranch.GetName(), nil
-}
-
-// GetPendingChanges returns the status of pending changes in the working directory
-func (e *engineImpl) GetPendingChanges(ctx context.Context) ([]PendingChange, error) {
-	output, err := e.git.GetStatusPorcelain(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	var changes []PendingChange
-	lines := strings.Split(strings.TrimSpace(output), "\n")
-	for _, line := range lines {
-		if len(line) < 4 {
-			continue
-		}
-		// Porcelain format: XY path
-		// X is staged status, Y is unstaged status
-		x := line[0]
-		y := line[1]
-		path := strings.TrimSpace(line[3:])
-
-		if x != ' ' && x != '?' {
-			changes = append(changes, PendingChange{
-				Path:   path,
-				Status: string(x),
-				Staged: true,
-			})
-		}
-		if y != ' ' {
-			status := string(y)
-			if x == '?' && y == '?' {
-				status = "??"
-			}
-			changes = append(changes, PendingChange{
-				Path:   path,
-				Status: status,
-				Staged: false,
-			})
-		}
-	}
-
-	return changes, nil
 }
 
 // Trunk returns the trunk branch
@@ -201,307 +155,6 @@ func (e *engineImpl) GetRelativeStack(branch Branch, rng StackRange) []Branch {
 	return result
 }
 
-// IsTrunk checks if a branch is the trunk
-func (e *engineImpl) IsTrunk(branch Branch) bool {
-	branchName := branch.GetName()
-	e.mu.RLock()
-	defer e.mu.RUnlock()
-	return branchName == e.trunk
-}
-
-// IsTracked checks if a branch is tracked (has metadata)
-func (e *engineImpl) IsTracked(branch Branch) bool {
-	branchName := branch.GetName()
-	e.mu.RLock()
-	defer e.mu.RUnlock()
-	_, ok := e.parentMap[branchName]
-	return ok
-}
-
-// GetScope returns the scope for a branch, inheriting from parent if not set
-func (e *engineImpl) GetScope(branch Branch) Scope {
-	branchName := branch.GetName()
-	e.mu.RLock()
-	defer e.mu.RUnlock()
-
-	current := branchName
-	for {
-		if scopeStr, ok := e.scopeMap[current]; ok && scopeStr != "" {
-			scope := NewScope(scopeStr)
-			if scope.IsNone() {
-				return Empty()
-			}
-			return scope
-		}
-		parent, ok := e.parentMap[current]
-		if !ok || parent == e.trunk {
-			break
-		}
-		current = parent
-	}
-	return Empty()
-}
-
-// IsLocked checks if a branch is locked
-func (e *engineImpl) IsLocked(branch Branch) bool {
-	e.mu.RLock()
-	defer e.mu.RUnlock()
-	return LockReason(e.lockedMap[branch.GetName()]).IsLocked()
-}
-
-// GetLockReason returns the reason why a branch is locked
-func (e *engineImpl) GetLockReason(branch Branch) LockReason {
-	e.mu.RLock()
-	defer e.mu.RUnlock()
-	return LockReason(e.lockedMap[branch.GetName()])
-}
-
-// IsFrozen checks if a branch is frozen
-func (e *engineImpl) IsFrozen(branch Branch) bool {
-	e.mu.RLock()
-	defer e.mu.RUnlock()
-	return e.frozenMap[branch.GetName()]
-}
-
-// GetExplicitScope returns the explicit scope set for a branch (no inheritance)
-func (e *engineImpl) GetExplicitScope(branch Branch) Scope {
-	branchName := branch.GetName()
-	e.mu.RLock()
-	defer e.mu.RUnlock()
-
-	scopeStr := e.scopeMap[branchName]
-	if scopeStr == "" {
-		return Empty()
-	}
-	return NewScope(scopeStr)
-}
-
-// getExplicitScope is an internal method for Branch type
-func (e *engineImpl) getExplicitScope(branch Branch) Scope {
-	return e.GetExplicitScope(branch)
-}
-
-// IsUpToDate checks if a branch is up to date with its parent
-// A branch is up to date if its parent revision matches the stored parent revision
-func (e *engineImpl) IsUpToDate(branch Branch) bool {
-	branchName := branch.GetName()
-	if e.IsTrunk(branch) {
-		return true
-	}
-
-	e.mu.RLock()
-	parent, ok := e.parentMap[branchName]
-	e.mu.RUnlock()
-
-	if !ok {
-		return true // Not tracked, consider it fixed
-	}
-
-	// Get current parent revision
-	parentRev, err := e.git.GetRevision(parent)
-	if err != nil {
-		return false // Can't determine, assume needs restack
-	}
-
-	// Get stored parent revision from metadata
-	meta, err := e.git.ReadMetadata(branchName)
-	if err != nil {
-		return false // No metadata, assume needs restack
-	}
-
-	if meta.ParentBranchRevision == nil {
-		return false // No stored revision, needs restack
-	}
-
-	// Branch is fixed if stored revision matches current parent revision
-	return *meta.ParentBranchRevision == parentRev
-}
-
-// GetCommitDate returns the commit date for a branch
-func (e *engineImpl) GetCommitDate(branch Branch) (time.Time, error) {
-	branchName := branch.GetName()
-	return e.git.GetCommitDate(branchName)
-}
-
-// GetCommitAuthor returns the commit author for a branch
-func (e *engineImpl) GetCommitAuthor(branch Branch) (string, error) {
-	branchName := branch.GetName()
-	return e.git.GetCommitAuthor(branchName)
-}
-
-// GetRevision returns the SHA of a branch
-func (e *engineImpl) GetRevision(branch Branch) (string, error) {
-	branchName := branch.GetName()
-	return e.git.GetRevision(branchName)
-}
-
-// GetCommitCount returns the number of commits for a branch
-func (e *engineImpl) GetCommitCount(branch Branch) (int, error) {
-	branchName := branch.GetName()
-	e.mu.RLock()
-	trunk := e.trunk
-	parent, ok := e.parentMap[branchName]
-	e.mu.RUnlock()
-
-	if !ok {
-		parent = trunk
-	}
-
-	// Get base revision (stored parent revision)
-	meta, err := e.git.ReadMetadata(branchName)
-	var base string
-	if err == nil && meta.ParentBranchRevision != nil {
-		base = *meta.ParentBranchRevision
-	} else {
-		// Fallback to current parent branch tip if metadata is missing
-		baseRev, err := e.git.GetRevision(parent)
-		if err != nil {
-			return 0, err
-		}
-		base = baseRev
-	}
-
-	branchRev, err := e.git.GetRevision(branchName)
-	if err != nil {
-		return 0, err
-	}
-
-	// If revisions are same, count is 0
-	if branchRev == base {
-		return 0, nil
-	}
-
-	// For real git, we'd use a git helper. I'll use git.GetCommitRange count.
-
-	commits, err := e.GetAllCommits(branch, CommitFormatSHA)
-	if err != nil {
-		return 0, err
-	}
-	return len(commits), nil
-}
-
-// GetDiffStats returns diff stats for a branch
-func (e *engineImpl) GetDiffStats(branch Branch) (int, int, error) {
-	branchName := branch.GetName()
-	e.mu.RLock()
-	trunk := e.trunk
-	parent, ok := e.parentMap[branchName]
-	e.mu.RUnlock()
-
-	if !ok {
-		parent = trunk
-	}
-
-	// Get base revision (stored parent revision)
-	meta, err := e.git.ReadMetadata(branchName)
-	var base string
-	if err == nil && meta.ParentBranchRevision != nil {
-		base = *meta.ParentBranchRevision
-	} else {
-		baseRev, err := e.git.GetRevision(parent)
-		if err != nil {
-			return 0, 0, err
-		}
-		base = baseRev
-	}
-
-	branchRev, err := e.git.GetRevision(branchName)
-	if err != nil {
-		return 0, 0, err
-	}
-
-	// If revisions are same, stats are 0
-	if branchRev == base {
-		return 0, 0, nil
-	}
-
-	// Use git diff --numstat
-	output, err := e.git.GetDiffNumstat(base, branchRev)
-	if err != nil {
-		return 0, 0, err
-	}
-
-	added, deleted := 0, 0
-	lines := strings.Split(strings.TrimSpace(output), "\n")
-	for _, line := range lines {
-		if line == "" {
-			continue
-		}
-		fields := strings.Fields(line)
-		if len(fields) >= 2 {
-			var a, d int
-			_, _ = fmt.Sscanf(fields[0], "%d", &a)
-			_, _ = fmt.Sscanf(fields[1], "%d", &d)
-			added += a
-			deleted += d
-		}
-	}
-
-	return added, deleted, nil
-}
-
-// BranchMatchesRemote checks if a branch matches its remote
-func (e *engineImpl) BranchMatchesRemote(branchName string) (bool, error) {
-	e.mu.RLock()
-	defer e.mu.RUnlock()
-
-	// First try to get remote SHA from cache (populated by PopulateRemoteShas)
-	remoteSha, exists := e.remoteShas[branchName]
-	if exists {
-		// Get local branch SHA
-		localSha, err := e.git.GetRevision(branchName)
-		if err != nil {
-			return false, nil
-		}
-		return localSha == remoteSha, nil
-	}
-
-	// Fall back to checking local remote tracking branch (like getBranchRemoteDifference does)
-	// This handles cases where remote fetching failed but we have local remote tracking
-	remoteTrackingSha, err := e.git.GetRemoteRevision(branchName)
-	if err != nil {
-		// No remote tracking branch exists
-		return false, nil
-	}
-
-	// Get local branch SHA
-	localSha, err := e.git.GetRevision(branchName)
-	if err != nil {
-		return false, nil
-	}
-
-	return localSha == remoteTrackingSha, nil
-}
-
-// IsMergedIntoTrunk checks if a branch is merged into trunk
-func (e *engineImpl) IsMergedIntoTrunk(ctx context.Context, branchName string) (bool, error) {
-	e.mu.RLock()
-	trunk := e.trunk
-	e.mu.RUnlock()
-	return e.git.IsMerged(ctx, branchName, trunk)
-}
-
-// IsBranchEmpty checks if a branch has no changes compared to its parent
-func (e *engineImpl) IsBranchEmpty(ctx context.Context, branchName string) (bool, error) {
-	e.mu.RLock()
-	parent, ok := e.parentMap[branchName]
-	trunk := e.trunk
-	e.mu.RUnlock()
-
-	if !ok {
-		// If not tracked, compare to trunk
-		parent = trunk
-	}
-
-	// Get parent revision
-	parentRev, err := e.git.GetRevision(parent)
-	if err != nil {
-		return false, err
-	}
-
-	return e.git.IsDiffEmpty(ctx, branchName, parentRev)
-}
-
 // FindMostRecentTrackedAncestors finds the most recent tracked ancestors of a branch
 // by checking the branch's commit history against tracked branch tips.
 // Returns a slice of branch names that point to the most recent tracked commit in history.
@@ -585,66 +238,6 @@ func (e *engineImpl) FindBranchForCommit(commitSHA string) (string, error) {
 	}
 
 	return "", nil
-}
-
-// GetAllCommits returns commits for a branch in various formats
-func (e *engineImpl) GetAllCommits(branch Branch, format CommitFormat) ([]string, error) {
-	branchName := branch.GetName()
-	e.mu.RLock()
-	defer e.mu.RUnlock()
-
-	// Check if branch is trunk
-	if branchName == e.trunk {
-		// Trunk is the base, so it has no commits "on" it relative to a parent
-		return []string{}, nil
-	}
-
-	// Get metadata to find parent revision
-	meta, err := e.git.ReadMetadata(branchName)
-	if err != nil {
-		return nil, err
-	}
-
-	// Get branch revision
-	branchRevision, err := e.git.GetRevision(branchName)
-	if err != nil {
-		return nil, err
-	}
-
-	// Get parent revision (base)
-	var baseRevision string
-	if meta.ParentBranchRevision != nil {
-		baseRevision = *meta.ParentBranchRevision
-	}
-
-	// Get SHAs first
-	shas, err := e.git.GetCommitRangeSHAs(baseRevision, branchRevision)
-	if err != nil {
-		return nil, err
-	}
-
-	if format == CommitFormatSHA {
-		return shas, nil
-	}
-
-	// Format results
-	result := make([]string, len(shas))
-	for i, sha := range shas {
-		var formatted string
-		switch format {
-		case CommitFormatSubject:
-			formatted, _ = e.git.GetCommitLog(sha, "%s")
-		case CommitFormatMessage:
-			formatted, _ = e.git.GetCommitLog(sha, "%B")
-		case CommitFormatReadable:
-			formatted, _ = e.git.GetCommitLog(sha, "%h %s")
-		default:
-			return nil, fmt.Errorf("unknown commit format: %s", format)
-		}
-		result[i] = strings.TrimSpace(formatted)
-	}
-
-	return result, nil
 }
 
 // GetRelativeStackUpstack returns all branches in the upstack (descendants)
@@ -741,46 +334,6 @@ func (e *engineImpl) SortBranchesTopologically(branches []Branch) []Branch {
 	return result
 }
 
-// GetDeletionStatus checks if a branch should be deleted
-func (e *engineImpl) GetDeletionStatus(ctx context.Context, branchName string) (DeletionStatus, error) {
-	// Check PR info
-	branch := e.GetBranch(branchName)
-	prInfo, err := e.GetPrInfo(branch)
-	if err == nil && prInfo != nil {
-		const (
-			prStateClosed = "CLOSED"
-			prStateMerged = "MERGED"
-		)
-		if prInfo.State() == prStateClosed {
-			return DeletionStatus{SafeToDelete: true, Reason: fmt.Sprintf("%s is closed on GitHub", branchName)}, nil
-		}
-		if prInfo.State() == prStateMerged {
-			base := prInfo.Base()
-			if base == "" {
-				base = e.Trunk().GetName()
-			}
-			return DeletionStatus{SafeToDelete: true, Reason: fmt.Sprintf("%s is merged into %s", branchName, base)}, nil
-		}
-	}
-
-	// Check if merged into trunk
-	merged, err := e.IsMergedIntoTrunk(ctx, branchName)
-	if err == nil && merged {
-		return DeletionStatus{SafeToDelete: true, Reason: fmt.Sprintf("%s is merged into %s", branchName, e.Trunk().GetName())}, nil
-	}
-
-	// Check if empty
-	empty, err := e.IsBranchEmpty(ctx, branchName)
-	if err == nil && empty {
-		// Only delete empty branches if they have a PR
-		if prInfo != nil && prInfo.Number() != nil {
-			return DeletionStatus{SafeToDelete: true, Reason: fmt.Sprintf("%s is empty", branchName)}, nil
-		}
-	}
-
-	return DeletionStatus{SafeToDelete: false, Reason: ""}, nil
-}
-
 // BranchesDepthFirst returns an iterator that yields branches starting from startBranch in depth-first order.
 // Each iteration yields (branchName, depth) where depth is 0 for the start branch.
 // The iterator can be used with range loops and supports early termination with break.
@@ -809,172 +362,4 @@ func (e *engineImpl) BranchesDepthFirst(startBranch Branch) iter.Seq2[Branch, in
 
 		visit(startBranch.GetName(), 0)
 	}
-}
-
-// GetRemote returns the default remote name
-func (e *engineImpl) GetRemote() string {
-	return e.git.GetRemote()
-}
-
-// GetBranchRemoteDifference returns a string describing the difference between local and remote branch
-func (e *engineImpl) GetBranchRemoteDifference(branchName string) (string, error) {
-	localSha, err := e.git.GetRevision(branchName)
-	if err != nil {
-		return "", fmt.Errorf("failed to get local SHA for %s: %w", branchName, err)
-	}
-
-	remoteSha, err := e.git.GetRemoteRevision(branchName)
-	if err != nil {
-		remote := e.git.GetRemote()
-		remoteShas, err := e.git.FetchRemoteShas(remote)
-		if err != nil {
-			localShort := localSha
-			if len(localSha) > 7 {
-				localShort = localSha[:7]
-			}
-			return fmt.Sprintf("local: %s (unable to fetch remote SHA)", localShort), nil
-		}
-		var exists bool
-		remoteSha, exists = remoteShas[branchName]
-		if !exists {
-			localShort := localSha
-			if len(localSha) > 7 {
-				localShort = localSha[:7]
-			}
-			return fmt.Sprintf("local: %s (branch not found on remote)", localShort), nil
-		}
-	}
-
-	if localSha == remoteSha {
-		return "", nil
-	}
-
-	localShort := localSha
-	if len(localSha) > 7 {
-		localShort = localSha[:7]
-	}
-	remoteShort := remoteSha
-	if len(remoteSha) > 7 {
-		remoteShort = remoteSha[:7]
-	}
-
-	remote := e.git.GetRemote()
-	remoteBranchRef := "refs/remotes/" + remote + "/" + branchName
-	commonAncestor, err := e.git.GetMergeBaseByRef(branchName, remoteBranchRef)
-	if err != nil {
-		return fmt.Sprintf("local: %s, remote: %s (likely local is ahead)", localShort, remoteShort), nil //nolint:nilerr
-	}
-
-	switch {
-	case commonAncestor == localSha:
-		return fmt.Sprintf("local is behind remote (local: %s, remote: %s)", localShort, remoteShort), nil
-	case commonAncestor == remoteSha:
-		return fmt.Sprintf("local is ahead of remote (local: %s, remote: %s)", localShort, remoteShort), nil
-	default:
-		return fmt.Sprintf("local and remote have diverged (local: %s, remote: %s)", localShort, remoteShort), nil
-	}
-}
-
-// HasStagedChanges checks if there are staged changes in the repository
-func (e *engineImpl) GetUnstagedDiff(ctx context.Context, files ...string) (string, error) {
-	return e.git.GetUnstagedDiff(ctx, files...)
-}
-
-func (e *engineImpl) HasStagedChanges(ctx context.Context) (bool, error) {
-	return e.git.HasStagedChanges(ctx)
-}
-
-// HasUnstagedChanges checks if there are unstaged changes in the repository
-func (e *engineImpl) HasUnstagedChanges(ctx context.Context) (bool, error) {
-	return e.git.HasUnstagedChanges(ctx)
-}
-
-// GetMergeBase returns the merge base between two revisions
-func (e *engineImpl) GetMergeBase(rev1, rev2 string) (string, error) {
-	return e.git.GetMergeBase(rev1, rev2)
-}
-
-// GetChangedFiles returns the list of files changed between base and head
-func (e *engineImpl) GetChangedFiles(ctx context.Context, base, head string) ([]string, error) {
-	return e.git.GetChangedFiles(ctx, base, head)
-}
-
-// ListWorktrees returns a list of worktree paths
-func (e *engineImpl) ListWorktrees(ctx context.Context) ([]string, error) {
-	return e.git.ListWorktrees(ctx)
-}
-
-// GetRemoteURL returns the remote URL
-func (e *engineImpl) GetRemoteURL(_ context.Context) (string, error) {
-	return e.git.GetConfig("remote.origin.url")
-}
-
-// GetCurrentRevision returns the current revision (HEAD)
-func (e *engineImpl) GetCurrentRevision(_ context.Context) (string, error) {
-	return e.git.GetRevision("HEAD")
-}
-
-// GetReflog returns the reflog
-func (e *engineImpl) GetReflog(ctx context.Context, count int, format string) (string, error) {
-	return e.git.GetReflog(ctx, count, format)
-}
-
-// CheckoutPaths checks out specific paths from a branch
-func (e *engineImpl) CheckoutPaths(ctx context.Context, branch string, pathspecs []string) error {
-	return e.git.CheckoutPaths(ctx, branch, pathspecs)
-}
-
-// RemovePaths removes specific paths from the working tree
-func (e *engineImpl) RemovePaths(ctx context.Context, pathspecs []string) error {
-	return e.git.RemovePaths(ctx, pathspecs)
-}
-
-// StashList returns the stash list
-func (e *engineImpl) StashList(ctx context.Context) (string, error) {
-	return e.git.ListStash(ctx)
-}
-
-// ParseStagedHunks parses the output of `git diff --cached` into structured hunks
-func (e *engineImpl) ParseStagedHunks(ctx context.Context) ([]git.Hunk, error) {
-	return e.git.ParseStagedHunks(ctx)
-}
-
-// ShowDiff returns the diff between two refs with optional stat mode
-func (e *engineImpl) ShowDiff(ctx context.Context, left, right string, stat bool) (string, error) {
-	return e.git.ShowDiff(ctx, left, right, stat)
-}
-
-// ShowCommits returns commit log with optional patches/stat
-func (e *engineImpl) ShowCommits(ctx context.Context, base, head string, patch, stat bool) (string, error) {
-	return e.git.ShowCommits(ctx, base, head, patch, stat)
-}
-
-// GetCommitTemplate returns the commit template
-func (e *engineImpl) GetCommitTemplate(ctx context.Context) (string, error) {
-	return e.git.GetCommitTemplate(ctx)
-}
-
-// GetUnmergedFiles returns list of files with merge conflicts
-func (e *engineImpl) GetUnmergedFiles(ctx context.Context) ([]string, error) {
-	return e.git.GetUnmergedFiles(ctx)
-}
-
-// GetParentCommitSHA returns the parent commit SHA of a commit
-func (e *engineImpl) GetParentCommitSHA(commitSHA string) (string, error) {
-	return e.git.GetParentCommitSHA(commitSHA)
-}
-
-// GetCommitSHA returns the SHA at a relative position (0 = HEAD, 1 = HEAD~1)
-func (e *engineImpl) GetCommitSHA(branchName string, offset int) (string, error) {
-	return e.git.GetCommitSHA(branchName, offset)
-}
-
-// IsAncestor checks if one commit is an ancestor of another
-func (e *engineImpl) IsAncestor(ancestor, descendant string) (bool, error) {
-	return e.git.IsAncestor(ancestor, descendant)
-}
-
-// IsRebaseInProgress checks if a rebase is in progress
-func (e *engineImpl) IsRebaseInProgress(ctx context.Context) bool {
-	return e.git.IsRebaseInProgress(ctx)
 }
