+++ internal/git/cherry_pick.go
@@ -7,16 +7,16 @@ import (
 )
 
 func (r *runner) CherryPick(ctx context.Context, commitSHA, onto string) (string, error) {
-	if _, err := r.runGitCommandWithContextInternal(ctx, "checkout", "--detach", onto); err != nil {
+	if _, err := r.RunGitCommandWithContext(ctx, "checkout", "--detach", onto); err != nil {
 		return "", fmt.Errorf("failed to checkout %s: %w", onto, err)
 	}
 
-	if _, err := r.runGitCommandWithContextInternal(ctx, "cherry-pick", commitSHA); err != nil {
-		_, _ = r.runGitCommandWithContextInternal(ctx, "cherry-pick", "--abort")
+	if _, err := r.RunGitCommandWithContext(ctx, "cherry-pick", commitSHA); err != nil {
+		_, _ = r.RunGitCommandWithContext(ctx, "cherry-pick", "--abort")
 		return "", fmt.Errorf("failed to cherry-pick %s: %w", commitSHA, err)
 	}
 
-	newSHA, err := r.runGitCommandWithContextInternal(ctx, "rev-parse", "HEAD")
+	newSHA, err := r.RunGitCommandWithContext(ctx, "rev-parse", "HEAD")
 	if err != nil {
 		return "", fmt.Errorf("failed to get new SHA after cherry-pick: %w", err)
 	}
@@ -25,11 +25,11 @@ func (r *runner) CherryPick(ctx context.Context, commitSHA, onto string) (string
 }
 
 func (r *runner) CherryPickSimple(ctx context.Context, commitSHA string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "cherry-pick", commitSHA)
+	_, err := r.RunGitCommandWithContext(ctx, "cherry-pick", commitSHA)
 	return err
 }
 
 func (r *runner) CherryPickAbort(ctx context.Context) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "cherry-pick", "--abort")
+	_, err := r.RunGitCommandWithContext(ctx, "cherry-pick", "--abort")
 	return err
 }
