+++ internal/git/reset.go
@@ -6,7 +6,7 @@ import (
 )
 
 func (r *runner) ResetMerge(ctx context.Context, revision string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "reset", "--merge", revision)
+	_, err := r.RunGitCommandWithContext(ctx, "reset", "--merge", revision)
 	if err != nil {
 		return fmt.Errorf("failed to reset --merge to %s: %w", revision, err)
 	}
@@ -14,7 +14,7 @@ func (r *runner) ResetMerge(ctx context.Context, revision string) error {
 }
 
 func (r *runner) HardReset(ctx context.Context, revision string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "reset", "--hard", revision)
+	_, err := r.RunGitCommandWithContext(ctx, "reset", "--hard", revision)
 	if err != nil {
 		return fmt.Errorf("failed to hard reset to %s: %w", revision, err)
 	}
@@ -22,7 +22,7 @@ func (r *runner) HardReset(ctx context.Context, revision string) error {
 }
 
 func (r *runner) SoftReset(ctx context.Context, revision string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "reset", "-q", "--soft", revision)
+	_, err := r.RunGitCommandWithContext(ctx, "reset", "-q", "--soft", revision)
 	if err != nil {
 		return fmt.Errorf("failed to soft reset to %s: %w", revision, err)
 	}
@@ -30,6 +30,6 @@ func (r *runner) SoftReset(ctx context.Context, revision string) error {
 }
 
 func (r *runner) MixedReset(ctx context.Context, revision string) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "reset", revision)
+	_, err := r.RunGitCommandWithContext(ctx, "reset", revision)
 	return err
 }
