+++ internal/git/staging.go
@@ -7,7 +7,7 @@ import (
 )
 
 func (r *runner) StageAll(ctx context.Context) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "add", "-A")
+	_, err := r.RunGitCommandWithContext(ctx, "add", "-A")
 	if err != nil {
 		return fmt.Errorf("failed to stage all changes: %w", err)
 	}
@@ -19,7 +19,7 @@ func (r *runner) StagePatch(_ context.Context) error {
 }
 
 func (r *runner) StageTracked(ctx context.Context) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "add", "-u")
+	_, err := r.RunGitCommandWithContext(ctx, "add", "-u")
 	if err != nil {
 		return fmt.Errorf("failed to stage tracked changes: %w", err)
 	}
@@ -40,7 +40,7 @@ func (r *runner) StageChanges(ctx context.Context, opts StagingOptions) error {
 	}
 
 	if opts.Update {
-		_, err := r.runGitCommandWithContextInternal(ctx, "add", "-u")
+		_, err := r.RunGitCommandWithContext(ctx, "add", "-u")
 		return err
 	}
 
@@ -48,7 +48,7 @@ func (r *runner) StageChanges(ctx context.Context, opts StagingOptions) error {
 }
 
 func (r *runner) HasStagedChanges(ctx context.Context) (bool, error) {
-	output, err := r.runGitCommandWithContextInternal(ctx, "diff", "--cached", "--shortstat")
+	output, err := r.RunGitCommandWithContext(ctx, "diff", "--cached", "--shortstat")
 	if err != nil {
 		return false, fmt.Errorf("failed to check staged changes: %w", err)
 	}
@@ -58,7 +58,7 @@ func (r *runner) HasStagedChanges(ctx context.Context) (bool, error) {
 func (r *runner) HasUnstagedChanges(ctx context.Context) (bool, error) {
 	// Use git diff to check for unstaged changes to tracked files
 	// This is more reliable than parsing porcelain output which gets trimmed
-	output, err := r.runGitCommandWithContextInternal(ctx, "diff", "--name-only")
+	output, err := r.RunGitCommandWithContext(ctx, "diff", "--name-only")
 	if err != nil {
 		return false, fmt.Errorf("failed to check unstaged changes: %w", err)
 	}
@@ -66,7 +66,7 @@ func (r *runner) HasUnstagedChanges(ctx context.Context) (bool, error) {
 }
 
 func (r *runner) HasUntrackedFiles(ctx context.Context) (bool, error) {
-	output, err := r.runGitCommandWithContextInternal(ctx, "ls-files", "--others", "--exclude-standard")
+	output, err := r.RunGitCommandWithContext(ctx, "ls-files", "--others", "--exclude-standard")
 	if err != nil {
 		return false, fmt.Errorf("failed to check for untracked files: %w", err)
 	}
@@ -74,7 +74,7 @@ func (r *runner) HasUntrackedFiles(ctx context.Context) (bool, error) {
 }
 
 func (r *runner) ParseStagedHunks(ctx context.Context) ([]Hunk, error) {
-	diffOutput, err := r.runGitCommandRawWithContextInternal(ctx, "diff", "--cached")
+	diffOutput, err := r.RunGitCommandRawWithContext(ctx, "diff", "--cached")
 	if err != nil {
 		return nil, fmt.Errorf("failed to get staged diff: %w", err)
 	}
