+++ internal/git/stash.go
@@ -10,7 +10,7 @@ func (r *runner) StashPush(ctx context.Context, message string) (string, error) 
 	if message != "" {
 		args = append(args, "-m", message)
 	}
-	output, err := r.runGitCommandWithContextInternal(ctx, args...)
+	output, err := r.RunGitCommandWithContext(ctx, args...)
 	if err != nil {
 		return "", fmt.Errorf("stash push failed: %w", err)
 	}
@@ -18,7 +18,7 @@ func (r *runner) StashPush(ctx context.Context, message string) (string, error) 
 }
 
 func (r *runner) StashPop(ctx context.Context) error {
-	_, err := r.runGitCommandWithContextInternal(ctx, "stash", "pop")
+	_, err := r.RunGitCommandWithContext(ctx, "stash", "pop")
 	if err != nil {
 		return fmt.Errorf("stash pop failed: %w", err)
 	}
@@ -26,5 +26,5 @@ func (r *runner) StashPop(ctx context.Context) error {
 }
 
 func (r *runner) ListStash(ctx context.Context) (string, error) {
-	return r.runGitCommandWithContextInternal(ctx, "stash", "list")
+	return r.RunGitCommandWithContext(ctx, "stash", "list")
 }
